requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
bleach==6.3.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
comm==0.2.3
cuda-bindings==13.2.0
cuda-pathfinder==1.5.0
cuda-toolkit==13.0.2
debugpy==1.8.20
decorator==5.2.1
deepframe==0.1.2
deeplake==4.5.9
dnspython==2.8.0
email-validator==2.3.0
executing==2.2.1
fastapi==0.135.2
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.1
fastar==0.9.0
filelock==3.25.2
fsspec==2026.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hub==3.0.1
idna==3.11
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.8.0
jupyter_core==5.9.1
kaggle==2.0.0
kagglehub==1.0.0
kagglesdk==0.1.16
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.3
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
packaging==26.0
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
prompt_toolkit==3.0.52
protobuf==7.34.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
python-slugify==8.0.4
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.33.0
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
sentry-sdk==2.56.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
starlette==1.0.0
sympy==1.13.1
text-unidecode==1.3
torch==2.11.0+cu126
torchvision==0.26.0+cu126
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.6.0
webencodings==0.5.1
websockets==16.0
